+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+- Added the IRA page with the EQIP section to display 2023 data and predicted 2024-2031 data with total dollar amount and practice options [#268](https://github.com/policy-design-lab/pdl-frontend/issues/268) 
+
 ## [0.20.0] - 2024-06-03
 
 ### Changed